+++ cypress/integration/v2/blocks/heading/desktop-advanced.spec.js
@@ -1,0 +1,11 @@
+/**
+ * External dependencies
+ */
+import { registerTests } from '~stackable-e2e/helpers'
+
+/**
+ * Internal dependencies
+ */
+import { desktopAdvanced } from '.'
+
+describe( 'Heading Block ( Desktop Advanced )', registerTests( [ desktopAdvanced ] ) )
