+++ cypress/integration/v2/blocks/heading/desktop-style.spec.js
@@ -1,0 +1,11 @@
+/**
+ * External dependencies
+ */
+import { registerTests } from '~stackable-e2e/helpers'
+
+/**
+ * Internal dependencies
+ */
+import { desktopStyle } from '.'
+
+describe( 'Heading Block ( Desktop Style )', registerTests( [ desktopStyle ] ) )
